requirements.txt
flask==1.0.2
sqlalchemy==1.2.12
flask-sqlalchemy==2.3.1
pyjwt==1.6.4
pymysql==0.9.2
pyopenssl==18.0.0
python-dotenv==0.9.1
requests==2.20.0
flask-jwt-extended==3.13.1
